requirements.txt
selenium>=4.15
webdriver-manager>=4.0
openai>=1.0
anthropic>=0.30
# FastAPI + server
fastapi>=0.100
uvicorn>=0.20
python-multipart>=0.0.7
# Database
sqlalchemy>=2.0
alembic>=1.13
# Auth
pyjwt>=2.9
bcrypt>=4.0
cryptography>=43.0
# Utilities
pydantic-settings>=2.0
httpx>=0.27
